+++ test/services/storage_services/XRootDStorageService/XRootDStorageServiceBasicFunctionalTest.cpp
@@ -250,6 +250,7 @@ void XRootDServiceBasicFunctionalTest::do_BasicFunctionality_test(std::string ar
 
     // Create a XRootD Manager object
     wrench::XRootD::Deployment xrootd_deployment(simulation, {{wrench::XRootD::Property::CACHE_MAX_LIFETIME, "28800"}, {wrench::XRootD::Property::REDUCED_SIMULATION, arg}, {wrench::XRootD::Property::FILE_NOT_FOUND_TIMEOUT, "10"}}, {{wrench::StorageServiceMessagePayload::FILE_WRITE_REQUEST_MESSAGE_PAYLOAD, 1024}});
+    auto s = xrootd_deployment.size(); // coverage
 
     this->root_supervisor = xrootd_deployment.createRootSupervisor("Host1");