+++ RELEASENOTES.md
@@ -3,6 +3,7 @@ WRENCH Release Notes
 ### wrench 2.9-dev
 
   - Implementation of `reclaimHosts()` and `releaseHosts()` methods for the batch compute service, by which one can make compute nodes temporarily (or permanently) unavailable at runtime at any time throughout the simulation.
+  - Added the possibility to start execution controllers dynamically
 
 ### wrench 2.8